More async tests
