+++ src/test/java/jslt2/Jslt2Test.java
@@ -458,110 +458,6 @@ public class Jslt2Test {
     }
     
     @Test
-    public void testAsyncConstant() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        input.set("name", TextNode.valueOf("tony"));
-        input.set("team", TextNode.valueOf("packers"));
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async.json").toPath()));         
-        JsonNode result = runtime.eval(script, input);
-        System.out.println(result);
-       // testAgainstSpec(input, query);
-        assertEquals("{\"name\":\"favre\",\"a\":true,\"b\":{\"b1\":\"z\",\"c1\":\"hi\",\"name\":\"tony\",\"team\":\"packers\"},\"c\":20,\"d\":[10,8],\"e\":{\"test\":{\"i\":0}},\"f\":2,\"g\":\"g2\"}", result.toString());
-    }
-    
-    @Test(expected=Jslt2Exception.class)
-    public void testAsyncBadFunctionReference() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-bad-def.json").toPath()));         
-        runtime.eval(script, input);
-    }
-    
-    @Test(expected=Jslt2Exception.class)
-    public void testAsyncBadVarReference() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-bad-var.json").toPath()));         
-        runtime.eval(script, input);
-    }
-    
-    @Test(expected=Jslt2Exception.class)
-    public void testAsyncReferenceVarInAsync() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-bad-var-in-async.json").toPath()));         
-        runtime.eval(script, input);        
-    }
-    
-    @Test(expected=Jslt2Exception.class)
-    public void testAsyncReferenceVarInAsync2() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-bad-var-in-async2.json").toPath()));         
-        JsonNode result = runtime.eval(script, input);
-        System.out.println(result);
-    }
-    
-    @Test
-    public void testAsyncReferenceVarInFunction() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-reference-var-in-function.json").toPath()));         
-        JsonNode result = runtime.eval(script, input);
-        System.out.println(result);
-        assertEquals("{\"k\":\"hi\"}", result.toString());
-    }
-    
-    @Test
-    public void testAsyncReferenceVar() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        
-        String script = new String(Files.readAllBytes(new File("./examples/async-reference-var.json").toPath()));         
-        JsonNode result = runtime.eval(script, input);
-        System.out.println(result);
-        assertEquals("{\"k\":\"hi\"}", result.toString());
-    }
-    
-    @Test
-    public void testAsyncEmbed() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();        
-        String script = new String(Files.readAllBytes(new File("./examples/async-embed.json").toPath()));         
-        long startTime = System.currentTimeMillis();
-        JsonNode result = runtime.eval(script, input);
-        long endTime = System.currentTimeMillis();
-        
-        System.out.println(result);       
-        System.out.println("Total time: " + (endTime - startTime) + " msec.");
-        assertEquals("{\"c\":\"2\"}", result.toString());
-     
-    }
-    
-    @Test
-    public void testAsyncSpeed() throws Exception {        
-        ObjectNode input = runtime.newObjectNode();
-        input.set("name", TextNode.valueOf("tony"));
-        input.set("team", TextNode.valueOf("packers"));
-        
-        String script = new String(Files.readAllBytes(new File("./examples/sleep.json").toPath()));      
-        long startTime = System.currentTimeMillis();
-        JsonNode result = runtime.eval(script, input);
-        long endTime = System.currentTimeMillis();
-        
-        System.out.println(result);       
-        System.out.println("(No Async) Total time: " + (endTime - startTime) + " msec.");
-        
-        
-        String asyncScript = new String(Files.readAllBytes(new File("./examples/sleep-async.json").toPath()));      
-        startTime = System.currentTimeMillis();
-        result = runtime.eval(asyncScript, input);
-        endTime = System.currentTimeMillis();
-        
-        System.out.println(result);       
-        System.out.println("(With Async) Total time: " + (endTime - startTime) + " msec.");
-    }
-        
-    @Test
     public void testArrays() throws Exception {        
         ArrayNode input = runtime.newArrayNode(12);
         for(int i = 0; i < 10; i++) {
